[AIRFLOW-3464] Move SkipMixin out of models.py (#4386)
